pyproject.toml
[project]
name = "prism"
version = "1.0.0"
description = "Clinical Trial Data Pipeline with AI-powered Code Generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.0.0",
    "pandas>=2.0.0",
    "polars>=1.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.10",
    "jinja2>=3.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
    "requests>=2.32.0",
    "python-dateutil>=2.8.0",
    "python-docx>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "httpcore>=1.0.0",
    "sniffio>=1.3.0",
    "socksio>=1.0.0",
    "zhipuai>=2.1.0",
    "mcp[cli]>=1.2.0",
]

[project.scripts]
prism = "prism.cli:main"
prism-mcp = "prism.mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prism"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
